+++ website/docusaurus.config.js
@@ -6,6 +6,13 @@ module.exports = {
   favicon: "img/favicon.ico",
   organizationName: "teslamate-org",
   projectName: "teslamate",
+  future: {
+    v4: true,
+  },
+  storage: {
+    type: 'localStorage',
+    namespace: true,
+ },
   themeConfig: {
     navbar: {
       title: "TeslaMate",